requirements.txt
fastapi
uvicorn
google-generativeai
gTTS
firebase-admin
pydub
google-cloud-storage
pillow
python-multipart